Move device specific code into device folder
