+++ src/devices/fujitsu/ar-ry13/checksum.js
