+++ src/devices/fujitsu/ar-ry13/payload.test.js
@@ -1,0 +1,47 @@
+import { describe, expect, it } from 'bun:test'
+import {
+  FANSPEED,
+  makeFujitsuPayload,
+  MODE,
+  SWING,
+  addProntoMetadata,
+} from './payload'
+
+describe('generating payload bytes', () => {
+  it('generates the expected payload', () => {
+    const output = makeFujitsuPayload(
+      23,
+      MODE.fan,
+      FANSPEED.high,
+      SWING.both,
+      false,
+    )
+
+    expect(output).toEqual([
+      0x28, 0xc6, 0x00, 0x08, 0x08, 0x7f, 0x90, 0x0c, 0x0e, 0x00, 0x8c, 0x00,
+      0x00, 0x00, 0x04, 0xf0,
+    ])
+  })
+})
+
+describe('encoding bytes into a pronto code', () => {
+  it('generates the expected pronto code', () => {
+    const input = [
+      0x28, 0xc6, 0x00, 0x08, 0x08, 0x7f, 0x90, 0x0c, 0x0e, 0x00, 0x8c, 0x00,
+      0x00, 0x00, 0x04, 0xf0,
+    ]
+
+    const output = addProntoMetadata(
+      input,
+      39e3,
+      [0x7c, 0x3e],
+      [0x10, 0x130],
+      [0x10, 0x2e],
+      [0x10, 0x10],
+    )
+
+    expect(output).toBe(
+      '0000006a00000082007c003e00100010001000100010002e001000100010002e0010001000100010001000100010002e0010002e0010001000100010001000100010002e0010002e001000100010001000100010001000100010001000100010001000100010001000100010001000100010001000100010001000100010002e001000100010001000100010001000100010001000100010001000100010002e001000100010001000100010001000100010002e0010002e0010002e0010002e0010002e0010002e0010002e0010002e00100010001000100010002e00100010001000100010001000100010001000100010001000100010001000100010002e0010002e0010001000100010001000100010001000100010001000100010002e0010002e0010002e0010001000100010001000100010001000100010001000100010001000100010001000100010002e0010001000100010001000100010002e0010002e001000100010001000100010001000100010001000100010001000100010001000100010001000100010001000100010001000100010001000100010001000100010001000100010001000100010001000100010001000100010001000100010001000100010001000100010001000100010001000100010001000100010002e00100010001000100010002e0010002e0010002e0010002e0010001000100010001000100010001000100130',
+    )
+  })
+})
